Update on QA angular details done

+++ apollo-api/src/main/kotlin/org/kebs/app/kotlin/apollo/api/handlers/QualityAssuranceInternalUserHandler.kt
@@ -86,7 +86,7 @@ class QualityAssuranceInternalUserHandler(
             val permitID =
                 req.paramOrNull("permitID")?.toLong() ?: throw ExpectedDataNotFound("Required Permit ID, check config")
 //            
-            qaDaoServices.findPermitDetails(permitID.toLong())
+            qaDaoServices.findPermitDetails(permitID)
                 .let {
                     ok().body(it)
                 }
